#978 Incomplete changes for bower support for plugins
